requirements.txt
Flask==3.1.0
gunicorn==23.0.0 
python-dotenv==1.1.0
ruff==0.11.4
psycopg2-binary==2.9.10
beautifulsoup4==4.13.4
requests==2.32.4
validators==0.35.0

